% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_announcement_msg.R
\name{create_announcement_msg}
\alias{create_announcement_msg}
\title{Create announcement message for lab meeting}
\usage{
create_announcement_msg(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=format_meeting_info]{format_meeting_info()}}}
}
\description{
Create announcement message for lab meeting
}
\examples{
create_announcement_msg()
create_announcement_msg("Sebastian Funk")
create_announcement_msg("Seb")
create_announcement_msg("Sebastian Funk", "someone")

}
