% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_announcement_reminder.R
\name{create_announcement_reminder}
\alias{create_announcement_reminder}
\title{Create announcement reminder for lab meeting}
\usage{
create_announcement_reminder(..., topic = "", time = "")
}
\arguments{
\item{...}{Arguments passed to \code{\link[=format_meeting_info]{format_meeting_info()}}}

\item{topic}{What will be the topic?}

\item{time}{Time of the meeting, in "HH:MM" format}
}
\description{
Create announcement reminder for lab meeting
}
\examples{
time <- format(lubridate::now() + lubridate::minutes(15), "\%H:\%M")
create_announcement_reminder(time = time)
create_announcement_reminder("Sebastian Funk", time = time)
create_announcement_reminder("Seb", time = time)
create_announcement_reminder("Sebastian Funk", "someone", time = time)
create_announcement_reminder(random = "Sebastian Funk", time = time)
create_announcement_reminder(
  chair = "Sebastian Funk",
  zoom_link = "https://example.com",
  time = time
)

}
