% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_coffee_reminder.R
\name{create_coffee_reminder}
\alias{create_coffee_reminder}
\title{Create a reminder for the weekly coffee}
\usage{
create_coffee_reminder(zoom_link = NULL)
}
\description{
Create a reminder for the weekly coffee
}
\examples{
create_coffee_reminder(zoom_link = "https://example.com")

}
