% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_meeting_info.R
\name{format_meeting_info}
\alias{format_meeting_info}
\title{Format meeting info as a simple bullet list}
\usage{
format_meeting_info(
  assignee = "",
  random = "",
  chair = "",
  topic = "",
  zoom_link = "",
  room = "",
  gsheet_id = Sys.getenv("GSHEET_ID")
)
}
\arguments{
\item{assignee}{Who is presenting this week? (one or more comma-separated
full names; "Everyone" maps to a channel-wide ping)}

\item{random}{Who is in charge of talking about a random interesting thing?}

\item{chair}{Who will chair this meeting?}

\item{topic}{Topic / kind of meeting (e.g. "Project updates", "Journal Club")}

\item{zoom_link}{What is the zoom link?}

\item{room}{What is the room?}

\item{gsheet_id}{The ID of the google sheet}
}
\description{
Format meeting info as a simple bullet list
}
\examples{
format_meeting_info("Sebastian Funk", topic = "Project updates")
format_meeting_info("Sebastian Funk, Liza Hadley", topic = "Project updates")
format_meeting_info(topic = "Coffee walk")

}
