% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meeting_info.R
\name{get_meeting_info}
\alias{get_meeting_info}
\title{Get info for this week's lab meeting}
\usage{
get_meeting_info(gsheet_id)
}
\arguments{
\item{gsheet_id}{The ID of the google sheet}
}
\description{
Get info for this week's lab meeting
}
\examples{
googlesheets4::gs4_deauth()
get_meeting_info(Sys.getenv("GSHEET_ID"))

}
