% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_user_ids.R
\name{update_user_ids}
\alias{update_user_ids}
\title{Update list of slack user IDs}
\usage{
update_user_ids(oauth_token = Sys.getenv("SLACKBOT_OAUTH"))
}
\arguments{
\item{oauth_token}{The OAuth token to access
\href{https://api.slack.com/methods/users.list}{slack users.list API endpoint}}
}
\description{
Update list of slack user IDs
}
