% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_announcement_msg.R
\name{create_announcement_msg}
\alias{create_announcement_msg}
\title{Create announcement message for lab meeting}
\usage{
create_announcement_msg(assignee)
}
\arguments{
\item{assignee}{Who is presenting this week?}
}
\description{
Create announcement message for lab meeting
}
\examples{
create_announcement_msg("Seb")

}
