% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_announcement_reminder.R
\name{create_announcement_reminder}
\alias{create_announcement_reminder}
\title{Create announcement reminder for lab meeting}
\usage{
create_announcement_reminder(
  assignee = NULL,
  topic = NULL,
  random = NULL,
  chair = NULL,
  zoom_link = NULL
)
}
\arguments{
\item{assignee}{Who is presenting this week?}

\item{topic}{What is the topic of this week's presentation?}

\item{random}{Who is charge of talking about a random interesting thing?}

\item{chair}{Who will chair this meeting?}

\item{zoom_link}{The link to the zoom room where the meeting is taking place}
}
\description{
Create announcement reminder for lab meeting
}
\examples{
create_announcement_reminder()
create_announcement_reminder("Seb")
create_announcement_reminder("Seb", "the number pi")
create_announcement_reminder("Seb", "the number pi", "someone")
create_announcement_reminder(random = "Seb")
create_announcement_reminder(chair = "Seb", zoom_link = "https://example.com")

}
