% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_meeting_info.R
\name{format_meeting_info}
\alias{format_meeting_info}
\title{Format meeting info as a simple bullet list}
\usage{
format_meeting_info(
  assignee = NULL,
  random = NULL,
  chair = NULL,
  oauth_token = Sys.getenv("SLACKBOT_OAUTH")
)
}
\arguments{
\item{assignee}{Who is presenting this week?}

\item{random}{Who is charge of talking about a random interesting thing?}

\item{chair}{Who will chair this meeting?}
}
\description{
Format meeting info as a simple bullet list
}
\examples{
format_meeting_info("Sebastian Funk")
format_meeting_info("Seb")
format_meeting_info("Sebastian Funk", "someone")
format_meeting_info(random = "Sebastian Funk")

}
