% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_id.R
\name{get_user_id}
\alias{get_user_id}
\title{Get slack user ID from their name}
\usage{
get_user_id(name, gsheet_id)
}
\arguments{
\item{name}{A character string with the name ("Firstname Lastname")}

\item{gsheet_id}{The ID of the google sheet}
}
\description{
Get slack user ID from their name
}
